+++ changelog.md
@@ -7,6 +7,7 @@
 - Added new rate-it component for passing custom rating items.
 - Moved glyphs file to root so it can be accessed by `import {fa_glyph} from `vue-rate-it/glyphs;` 
 - Added new "extendable" example.
+- Updated "Extending" section of docs.
 
 ## Version 1.0 (2017-04-26)
 - Initial release